Refactor - Assistant UseCases
