+++ Application Layer/UserServices/UserServices.BusinessLayer/UseCase/Assistant/Assistant - Course.cs
@@ -9,7 +9,7 @@ using OnlineServices.Common.RegistrationServices;
 using OnlineServices.Common.RegistrationServices.TransferObject;
 using OnlineServices.Common.Exceptions;
 
-namespace RegistrationServices.BusinessLayer.UseCase
+namespace RegistrationServices.BusinessLayer.UseCase.Assistant
 {
     public partial class AssistantRole : IRSAssistantRoleCourse
     {
@@ -45,7 +45,7 @@ namespace RegistrationServices.BusinessLayer.UseCase
 
         public List<CourseTO> GetCourses()
         {
-            return iRSUnitOfWork.CourseRepository.GetAll().Select(x=>x.ToDomain().ToTransfertObject()).ToList();
+            return iRSUnitOfWork.CourseRepository.GetAll().Select(x => x.ToDomain().ToTransfertObject()).ToList();
         }
 
         public bool RemoveCourse(CourseTO courseTo)
@@ -84,7 +84,7 @@ namespace RegistrationServices.BusinessLayer.UseCase
             }
 
             courseTo.Name.IsNullOrWhiteSpace("Missing Course Name");
-            
+
             try
             {
                 iRSUnitOfWork.CourseRepository.Update(courseTo.ToDomain().ToTransfertObject());
@@ -97,79 +97,4 @@ namespace RegistrationServices.BusinessLayer.UseCase
             }
         }
     }
-
-    //===================================================================================================================================
-    public partial class AssistantRole : IRSAssistantRoleSession
-    {
-        public bool AddSession(SessionTO sessionTO)
-        {
-            if (sessionTO is null)
-                throw new ArgumentNullException(nameof(sessionTO));
-
-            if (sessionTO.Id != 0)
-                throw new Exception("Existing Session");
-
-            try
-            {
-                iRSUnitOfWork.SessionRepository.Add(sessionTO.ToDomain().ToTransfertObject());
-
-                return true;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-        }
-
-        public bool UpdateSession(SessionTO sessionTO)
-        {
-            if (sessionTO is null)
-                throw new ArgumentNullException((nameof(sessionTO)));
-
-            if (sessionTO.Id == 0)
-                throw new Exception("User does not exist");
-
-            try
-            {
-                iRSUnitOfWork.SessionRepository.Update(sessionTO.ToDomain().ToTransfertObject());
-                return true;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-        }
-
-        public bool RemoveSession(SessionTO sessionTO)
-        {
-            if (sessionTO is null)
-                throw new ArgumentNullException(nameof(sessionTO));
-
-            if (sessionTO.Id == 0)
-                throw new Exception("User does not exist");
-
-            try
-            {
-                iRSUnitOfWork.SessionRepository.Remove(sessionTO.ToDomain().ToTransfertObject());
-
-                return true;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-
-        }
-
-        public List<SessionTO> GetSessions()
-        {
-            return iRSUnitOfWork.SessionRepository.GetAll().Select(x => x.ToDomain().ToTransfertObject()).ToList();
-        }
-
-        public SessionTO GetSessionById(int id)
-        {
-            return iRSUnitOfWork.SessionRepository.GetById(id);
-        }
-    }
-
 }
