--- Application Layer/UserServices/UserServices.BusinessLayerTests/UnitTest1.cs
@@ -1,13 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace RegistrationServices.BusinessLayerTests
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod()]
-        public void TestMethod1()
-        {
-        }
-    }
-}
