+++ Application Layer/UserServices/UserServices.BusinessLayerTests/UseCase/AssistantCourseTests/Assistant_AddCourseTest.cs
@@ -4,10 +4,11 @@ using Moq;
 using OnlineServices.Common.Exceptions;
 using OnlineServices.Common.RegistrationServices.TransferObject;
 using OnlineServices.Common.RegistrationServices.Interfaces;
-using RegistrationServices.BusinessLayer.UseCase;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using RegistrationServices.BusinessLayer.UseCase.Assistant;
+
 namespace RegistrationServices.BusinessLayerTests.UseCase
 {
     [TestClass]
