+++ Cross-Cutting/OnlineServices.Common/RegistrationServices/Interfaces/Repositories/IRSUserRepository.cs
